task3_retrieval_code/repository/requirements.txt
torch
torchvision
transformers>=4.40,<5
accelerate
faiss-cpu
numpy
pillow
matplotlib
tqdm
scikit-learn
medmnist
